[masa]: forcing fabs to long double

+++ src/masa_class.cpp
@@ -352,10 +352,10 @@ int MASA::manufactured_solution<Scalar>::poly_test()
   poly.set_coeffs( a );
 
   // Check to make sure we get back what we set
-  if( fabs( a0 - poly.get_coeffs( 0 ) ) > thresh ) return_flag = 1;
-  if( fabs( a1 - poly.get_coeffs( 1 ) ) > thresh ) return_flag = 1;
-  if( fabs( a2 - poly.get_coeffs( 2 ) ) > thresh ) return_flag = 1;
-  if( fabs( a3 - poly.get_coeffs( 3 ) ) > thresh ) return_flag = 1;
+  if( fabsl( a0 - poly.get_coeffs( 0 ) ) > thresh ) return_flag = 1;
+  if( fabsl( a1 - poly.get_coeffs( 1 ) ) > thresh ) return_flag = 1;
+  if( fabsl( a2 - poly.get_coeffs( 2 ) ) > thresh ) return_flag = 1;
+  if( fabsl( a3 - poly.get_coeffs( 3 ) ) > thresh ) return_flag = 1;
 
   // Check polynomial evaluation
   const Scalar x = 2.0;
@@ -365,7 +365,7 @@ int MASA::manufactured_solution<Scalar>::poly_test()
   computed_value = poly( x , &ierr);
   if(ierr != 0) return_flag=1;
 
-  if( fabs( exact_value - computed_value ) > thresh ) return_flag = 1;
+  if( fabsl( exact_value - computed_value ) > thresh ) return_flag = 1;
 
   // Check derivatives
   const Scalar dx = 62;
@@ -376,10 +376,10 @@ int MASA::manufactured_solution<Scalar>::poly_test()
 
   poly.eval_derivs( x, 4, derivs );
   
-  if( fabs( exact_value - derivs[0] ) > thresh ) return_flag = 1;
-  if( fabs( dx  - derivs[1] ) > thresh ) return_flag = 1;
-  if( fabs( d2x - derivs[2] ) > thresh ) return_flag = 1;
-  if( fabs( d3x - derivs[3] ) > thresh ) return_flag = 1;
+  if( fabsl( exact_value - derivs[0] ) > thresh ) return_flag = 1;
+  if( fabsl( dx  - derivs[1] ) > thresh ) return_flag = 1;
+  if( fabsl( d2x - derivs[2] ) > thresh ) return_flag = 1;
+  if( fabsl( d3x - derivs[3] ) > thresh ) return_flag = 1;
 
   return return_flag;
 }
